+++ src/preload/index.d.ts
@@ -4,10 +4,10 @@ declare global {
   interface Window {
     electron: ElectronAPI;
     api: {
-      nf: {
+      invoice: {
         list: () => Promise<unknown[]>;
-        create: (nf: unknown) => Promise<unknown>;
-        update: (nf: unknown) => Promise<unknown>;
+        create: (invoice: unknown) => Promise<unknown>;
+        update: (invoice: unknown) => Promise<unknown>;
         delete: (id: string) => Promise<void>;
       };
       print: {